requirements.txt
gymnasium-robotics[mujoco-py]
mujoco
stable-baselines3[extra]
sb3-contrib
streamlit
gym